Split migrations and model definition tests into another test suite
