+++ runtests.py
@@ -17,7 +17,7 @@ installed_apps = [
     'simple_history.tests',
     'simple_history.tests.custom_user',
     'simple_history.tests.external',
-    'simple_history.tests.migration_test_app',
+    'simple_history.registry_tests.migration_test_app',
 
     'simple_history',
 
@@ -51,13 +51,6 @@ DEFAULT_SETTINGS = dict(
 
 
 def main():
-    # reset the test app migrations
-    for migration_path in [
-        join(dirname(__file__), 'simple_history', 'tests', 'migrations'),
-        join(dirname(__file__), 'simple_history', 'tests', 'migration_test_app', 'migrations'),
-    ]:
-        for migration_file_path in listdir(migration_path):
-            unlink(join(migration_path, migration_file_path))
 
     if not settings.configured:
         settings.configure(**DEFAULT_SETTINGS)
@@ -68,9 +61,10 @@ def main():
     except ImportError:
         from django.test.simple import DjangoTestSuiteRunner
         failures = DjangoTestSuiteRunner(failfast=False).run_tests(['tests'])
+        failures |= DjangoTestSuiteRunner(failfast=False).run_tests(['registry_tests'])
     else:
-        failures = DiscoverRunner(failfast=False).run_tests(
-            ['simple_history.tests'])
+        failures = DiscoverRunner(failfast=False).run_tests(['simple_history.tests'])
+        failures |= DiscoverRunner(failfast=False).run_tests(['simple_history.registry_tests'])
     sys.exit(failures)
