+++ simple_history/tests/tests/test_models.py
@@ -10,19 +10,15 @@ from django.db.models.fields.proxy import OrderWrt
 from django.test import TestCase
 from django.core.files.base import ContentFile
 
-from simple_history import exceptions, register
 from simple_history.models import HistoricalRecords, convert_auto_field
 from ..models import (
-    AdminProfile, Bookcase, MultiOneToOne, Poll, Choice, Voter, Restaurant,
+    AdminProfile, Bookcase, MultiOneToOne, Poll, Choice, Restaurant,
     Person, FileModel, Document, Book, HistoricalPoll, Library, State,
     AbstractBase, ConcreteAttr, ConcreteUtil, SelfFK, Temperature, WaterLevel,
     ExternalModel1, ExternalModel3, UnicodeVerboseName, HistoricalChoice,
     HistoricalState, HistoricalCustomFKError, Series, SeriesWork, PollInfo,
-    UserAccessorDefault, UserAccessorOverride, Employee, Country, Province,
+    Employee, Country, Province,
     City, Contact, ContactRegister,
-    TrackedAbstractBaseA, TrackedAbstractBaseB,
-    TrackedWithAbstractBase, TrackedWithConcreteBase,
-    InheritTracking1, InheritTracking2, InheritTracking3, InheritTracking4,
 )
 from ..external.models import ExternalModel2, ExternalModel4
 
@@ -317,39 +313,6 @@ class HistoricalRecordsTest(TestCase):
         poll.save()
         poll_info = PollInfo(poll=poll)
         poll_info.save()
-
-
-class RegisterTest(TestCase):
-    def test_register_no_args(self):
-        self.assertEqual(len(Choice.history.all()), 0)
-        poll = Poll.objects.create(pub_date=today)
-        choice = Choice.objects.create(poll=poll, votes=0)
-        self.assertEqual(len(choice.history.all()), 1)
-
-    def test_register_separate_app(self):
-        def get_history(model):
-            return model.history
-
-        self.assertRaises(AttributeError, get_history, User)
-        self.assertEqual(len(User.histories.all()), 0)
-        user = User.objects.create(username='bob', password='pass')
-        self.assertEqual(len(User.histories.all()), 1)
-        self.assertEqual(len(user.histories.all()), 1)
-
-    def test_reregister(self):
-        with self.assertRaises(exceptions.MultipleRegistrationsError):
-            register(Restaurant, manager_name='again')
-
-    def test_register_custome_records(self):
-        self.assertEqual(len(Voter.history.all()), 0)
-        poll = Poll.objects.create(pub_date=today)
-        choice = Choice.objects.create(poll=poll, votes=0)
-        user = User.objects.create(username='voter')
-        voter = Voter.objects.create(choice=choice, user=user)
-        self.assertEqual(len(voter.history.all()), 1)
-        expected = 'Voter object changed by None as of '
-        self.assertEqual(expected,
-                         str(voter.history.all()[0])[:len(expected)])
 
 
 class CreateHistoryModelTests(unittest.TestCase):
@@ -738,18 +701,6 @@ class TestLatest(TestCase):
         assert HistoricalPoll.objects.latest().pk == 1
 
 
-class TestUserAccessor(unittest.TestCase):
-
-    def test_accessor_default(self):
-        register(UserAccessorDefault)
-        assert not hasattr(User, 'historicaluseraccessordefault_set')
-
-    def test_accessor_override(self):
-        register(UserAccessorOverride,
-                 user_related_name='my_history_model_accessor')
-        assert hasattr(User, 'my_history_model_accessor')
-
-
 class TestMissingOneToOne(TestCase):
 
     def setUp(self):
@@ -790,67 +741,3 @@ class CustomTableNameTest1(TestCase):
             self.get_table_name(ContactRegister.history),
             'contacts_register_history',
         )
-
-
-class TestTrackingInheritance(TestCase):
-
-    def test_tracked_abstract_base(self):
-        self.assertEqual(
-            [
-                f.attname
-                for f in TrackedWithAbstractBase.history.model._meta.fields
-            ],
-            [
-                'id', 'history_id', 'history_date', 'history_user_id',
-                'history_type',
-            ],
-        )
-
-    def test_tracked_concrete_base(self):
-        self.assertEqual(
-            [
-                f.attname
-                for f in TrackedWithConcreteBase.history.model._meta.fields
-            ],
-            [
-                'id', 'trackedconcretebase_ptr_id', 'history_id',
-                'history_date', 'history_user_id', 'history_type',
-            ],
-        )
-
-    def test_multiple_tracked_bases(self):
-        with self.assertRaises(exceptions.MultipleRegistrationsError):
-            class TrackedWithMultipleAbstractBases(
-                    TrackedAbstractBaseA, TrackedAbstractBaseB):
-                pass
-
-    def test_tracked_abstract_and_untracked_concrete_base(self):
-        self.assertEqual(
-            [f.attname for f in InheritTracking1.history.model._meta.fields],
-            [
-                'id', 'untrackedconcretebase_ptr_id', 'history_id',
-                'history_date', 'history_user_id', 'history_type',
-            ],
-        )
-
-    def test_indirect_tracked_abstract_base(self):
-        self.assertEqual(
-            [f.attname for f in InheritTracking2.history.model._meta.fields],
-            [
-                'id', 'baseinherittracking2_ptr_id', 'history_id',
-                'history_date', 'history_user_id', 'history_type',
-            ],
-        )
-
-    def test_indirect_tracked_concrete_base(self):
-        self.assertEqual(
-            [f.attname for f in InheritTracking3.history.model._meta.fields],
-            [
-                'id', 'baseinherittracking3_ptr_id', 'history_id',
-                'history_date', 'history_user_id', 'history_type',
-            ],
-        )
-
-    def test_registering_with_tracked_abstract_base(self):
-        with self.assertRaises(exceptions.MultipleRegistrationsError):
-            register(InheritTracking4)
